plant_classifier/requirements.txt
deeplake
tensorflow
flask
matplotlib
pillow
